+++ blockservice/test/mock.go
@@ -9,13 +9,13 @@ import (
 )
 
 // Mocks returns |n| connected mock Blockservices
-func Mocks(n int, opts ...blockservice.Option) []blockservice.BlockService {
+func Mocks(n int, opts ...blockservice.Option) []*blockservice.BlockService {
 	net := tn.VirtualNetwork(mockrouting.NewServer(), delay.Fixed(0))
 	sg := testinstance.NewTestInstanceGenerator(net, nil, nil)
 
 	instances := sg.Instances(n)
 
-	var servs []blockservice.BlockService
+	var servs []*blockservice.BlockService
 	for _, i := range instances {
 		servs = append(servs, blockservice.New(i.Blockstore(), i.Exchange, opts...))
 	}